feat/add popup for activated user

+++ frontend/src/service/forms/activate-user-page.tsx
@@ -1,32 +1,42 @@
-//@ts-nocheck
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
-import { useAuth } from "../authContext";
 import { API } from "../axios";
 import "../../css/app.scss";
+import "../../css/activateUser.scss";
 
 function ActivateUserPage() {
   let { userId, userToken } = useParams();
   const navigate = useNavigate();
-  const { login } = useAuth();
   const [activationStatus, setActivationStatus] = useState("pending"); // New state for activation status
+  const [showModal, setShowModal] = useState(false);
+  const [answer, setAnswer] = useState(false);
 
   useEffect(() => {
-    API.get(`confirm-email/${userId}/${userToken}/`)
-      .then((response) => {
-        if (response.data) {
-          if (window.location.pathname.includes("/activate-client/")) {
-            localStorage.setItem("refreshToken", response.data.refresh_token);
-            localStorage.setItem("accessToken", response.data.access_token);
-            navigate(`/client-registration`);
-          }
-        }
-      })
-      .catch((error) => {
-        console.error("Error activating your account", error);
-        setActivationStatus("failed");
-      });
+    if (localStorage.getItem("accessToken")) {
+      setShowModal(true);
+    } else {
+      setAnswer(true);
+    }
   }, []);
+
+  useEffect(() => {
+    if (answer) {
+      API.get(`confirm-email/${userId}/${userToken}/`)
+        .then((response) => {
+          if (response.data) {
+            if (window.location.pathname.includes("/activate-client/")) {
+              localStorage.setItem("refreshToken", response.data.refresh_token);
+              localStorage.setItem("accessToken", response.data.access_token);
+              navigate(`/client-registration`);
+            }
+          }
+        })
+        .catch((error) => {
+          console.error("Error activating your account", error);
+          setActivationStatus("failed");
+        });
+    }
+  }, [answer]);
 
   return (
     <div>
@@ -35,6 +45,33 @@ function ActivateUserPage() {
       )}
       {activationStatus === "failed" && (
         <h1>Account Activation Failed. Please try again.</h1>
+      )}
+      {showModal && (
+        <div className="modal-overlay">
+          <div className="modal">
+            <div className="modal-div">
+              <p>
+                It seems that you are trying to create another account. <br />
+                Would you like to continue as a client or stay logged into your
+                psychologist account?
+              </p>
+              <div className="container__button">
+                <button
+                  className="action-button"
+                  onClick={() => {
+                    localStorage.clear();
+                    setAnswer(true);
+                  }}
+                >
+                  Continue as Client
+                </button>
+                <button className="action-button" onClick={() => navigate("/")}>
+                  Stay as Psychologist
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
       )}
     </div>
   );
